# [WIP] Vectorized partition function for floats on avx2 machines.