--- a/aten/src/ATen/NumericUtils.h
+++ b/aten/src/ATen/NumericUtils.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <cmath>
 #include <type_traits>