--- a/aten/src/ATen/native/native_functions.yaml
+++ b/aten/src/ATen/native/native_functions.yaml
@@ -4661,6 +4661,7 @@
   matches_jit_signature: True
   python_module: nn
 
+
 - func: reflection_pad1d(Tensor self, int[2] padding, *, Tensor(a!) out) -> Tensor(a!)
   matches_jit_signature: True
   python_module: nn