Speeds up kthvalue on cpu; can be used for topk and sort in the future.

The idea is to compare 8 elements at a time to the pivot, then reshuffle them within a 256-bit register by using a lookup table according to the comparison, and then write them back into the array at the required positions. (I got most of the ideas from [here](http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.1009.7773&rep=rep1&type=pdf) and then modified the algorithm a bit to be in-place.)

Check the comments in `qs_partition_inplace` of `QuickPartition.cpp` for more details.

This can be easily extended to doubles, but the speedup might not be as great because you'll only sort 4 elements at a time.

Future possibilities:
* Port topk and sort to use the same logic
* Make it work for types other than float32
* Figure out why it is only slightly faster on a large, fast-moving dimension on multi-core (see the 1000x1000000 case below)

Timings:

| Experiment  | Before | After | Speedup |
| --- | --- | --- | --- |
| Many cores, 1000000x1000 floats, dim 0  |  24.1s | 3.27s | 7.4x |
| Many cores, 1000000x1000 floats, dim 1  | 3.62s | 577ms | 6.3x |
| Many cores, 1000x1000000 floats, dim 0 | 4.06s | 727ms | 5.6x |
| Many cores, 1000x1000000 floats, dim 1 | 3.92s | 3.4s | 1.15x |
| One core, 1000000x1000 floats, dim 0 | 1m7s | 26.9s | 2.5x |
| One core, 1000000x1000 floats, dim 1 | 18s | 7.54s | 2.4x |
| One core, 1000x1000000 floats, dim 0 | 27.2s | 9.29s | 2.9x |
| One core, 1000x1000000 floats, dim 1 | 14.5s | 4.1s | 3.5x |