@ezyang 2019-04-11
This PR is not complete because it needs to use the registration mechanism to decide whether or not to use the AVX2 kernel or not. That likely explains the test failures; I don't see any use of our existing dispatching mechanisms for AVX2 kernels.