@cpuhrsch 2019-04-11
Can't some of this stuff be done using TensorIterator reductions?