@ezyang 2019-04-11
General comment: I have a hard time telling what `largest` is supposed to do. I thought it lets you invert which way the sort goes, but there are a few places where this expectation feels violated. E.g., https://github.com/pytorch/pytorch/pull/18344/files#diff-625f0146fdeb14ca4158305031e2c5e8R303