@ezyang 2019-04-12
> General comment: I have a hard time telling what largest is supposed to do. I thought it lets you invert which way the sort goes, but there are a few places where this expectation feels violated.

I think the problem is that what I think of as the logical sense for `largest` is inverted in this PR. I would have expected `largest == true` when I am attempting to compute the kth largest element. `torch.kthvalue` is defined to return the kth smallest element. Thus, I expect `largest == false` in this PR.  However, in https://github.com/pytorch/pytorch/pull/18344/files#diff-dfc3d7c56800ce763e3798bc86d964dcR224 you have passed `largest = true`. So this seems to imply that the opposite sense is being used here.