@umanwizard 2020-01-03
It seems unlikely that this is ever going to land. Closing the PR.